+++ src/main/java/com/chrisrm/idea/utils/UIReplacer.java
@@ -1,9 +1,14 @@
 package com.chrisrm.idea.utils;
 
+import com.intellij.openapi.wm.impl.status.MemoryUsagePanel;
 import com.intellij.ui.Gray;
 import com.intellij.util.ui.UIUtil;
 
 import javax.swing.*;
+import java.awt.*;
+import java.lang.reflect.Field;
+import java.util.Arrays;
+import java.util.stream.Stream;
 
 public class UIReplacer {
     public static void patchUI() {
@@ -11,6 +16,7 @@ public class UIReplacer {
             Patcher.patchTables();
             Patcher.patchStatusBar();
             Patcher.patchPanels();
+            Patcher.patchMemoryIndicator();
         }
         catch (Exception e) {
             e.printStackTrace();
@@ -43,7 +49,19 @@ public class UIReplacer {
             StaticPatcher.setFinalStatic(UIUtil.class, "BORDER_COLOR", color);
             StaticPatcher.setFinalStatic(UIUtil.class, "AQUA_SEPARATOR_FOREGROUND_COLOR", color);
         }
+
+        public static void patchMemoryIndicator() throws Exception {
+            Object usedColor = UIManager.getColor("MemoryIndicator.usedColor");
+            Object unusedColor = UIManager.getColor("MemoryIndicator.unusedColor");
+            StaticPatcher.setFinalStatic(MemoryUsagePanel.class, "USED_COLOR", usedColor);
+            StaticPatcher.setFinalStatic(MemoryUsagePanel.class, "UNUSED_COLOR", unusedColor);
+
+            Field[] fields = MemoryUsagePanel.class.getDeclaredFields();
+            Object[] objects = Arrays.stream(fields)
+                    .filter(f -> f.getType().equals(Color.class))
+                    .toArray();
+            StaticPatcher.setFinalStatic((Field) objects[0], usedColor);
+            StaticPatcher.setFinalStatic((Field) objects[1], unusedColor);
+        }
     }
-
-
 }